requirements.txt
streamlit
pymongo
pandas
openai==0.28
python-dotenv
datasets
googlemaps
Flask